--- ts/beamstreams.ts
@@ -1,486 +1,0 @@
-/*
-  Aerostat Beam Coder - Node.js native bindings to FFmpeg
-  Copyright (C) 2019 Streampunk Media Ltd.
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-  https://www.streampunk.media/ mailto:furnace@streampunk.media
-  14 Ormiscaig, Aultbea, Achnasheen, IV22 2JJ  U.K.
-*/
-import { Writable, Readable, Transform } from 'stream';
-import type { ffStats } from './types';
-import { BalanceResult, teeBalancer } from './teeBalancer';
-import { localFrame, parallelBalancer } from './parallelBalancer';
-import { serialBalancer } from './serialBalancer';
-import { Demuxer } from './types/Demuxer';
-import { Muxer } from './types/Muxer';
-import { Stream } from './types/Stream';
-import { DecodedFrames } from './types/DecodedFrames';
-import type { Governor } from './types/Governor';
-import { Frame } from './types/Frame';
-import { Packet } from './types/Packet';
-import { Timing, TotalTimeed } from './types/Timing';
-import { BeamstreamChannel, BeamstreamParams, BeamstreamSource, BeamstreamStream, ReadableMuxerStream, WritableDemuxerStream } from './types/Beamstreams';
-import { Filterer, FiltererResult, FilterLink } from './types/Filter';
-import { Timable, Timables } from './types/Timable'
-import { EncodedPackets } from './types/Encoder';
-
-import beamcoder from './beamcoder'
-import frameDicer from './frameDicer';
-
-const doTimings = false;
-const timings = [] as Array<{ [key: string]: Timing }>;
-
-// SRC = Frame | Frame[] | Packet
-// DST = Promise<DecodedFrames>, DecodedFrames
-function transformStream<SRC extends Timable, DST extends Timable>(
-  params: { name: 'encode' | 'dice' | 'decode' | 'filter', highWaterMark: number },
-  processFn: (val: SRC) => DST,
-  flushFn: () => DST | null | void,
-  reject: (err?: Error) => void) {
-  return new Transform({
-    objectMode: true,
-    highWaterMark: params.highWaterMark ? params.highWaterMark || 4 : 4,
-    transform(val: SRC, encoding, cb) {
-      (async () => {
-        const start = process.hrtime();
-        const reqTime = start[0] * 1e3 + start[1] / 1e6;
-        const result = await processFn(val);
-        result.timings = val.timings;
-        if (result.timings)
-          result.timings[params.name] = { reqTime, elapsed: process.hrtime(start)[1] / 1000000 };
-        cb(null, result);
-      })().catch(cb);
-    },
-    flush(cb) {
-      (async () => {
-        const result = flushFn ? await flushFn() : null;
-        if (result) result.timings = {};
-        cb(null, result);
-      })().catch(cb);
-    }
-  }).on('error', err => reject(err));
-}
-
-function calcStats<K extends string, P extends string>(arr: Array<{ [key in K]: { [prop in P]: number } }>, elem: K, prop: P): ffStats {
-  const values: number[] = arr.filter(cur => cur[elem]).map(cur => cur[elem][prop]);
-  const mean: number = values.reduce((acc, cur) => acc + cur, 0) / arr.length;
-  const max: number = Math.max(...values)
-  const min: number = Math.min(...values)
-  // standard deviation
-  const sumDelta: number = values.reduce((acc, cur) => acc + Math.pow(cur - mean, 2), 0);
-  const stdDev: number = Math.pow(sumDelta / arr.length, 0.5);
-  return { mean, stdDev, max, min };
-};
-
-function writeStream<T extends Timable, R>(params: { name: string, highWaterMark?: number }, processFn: (val: T) => Promise<R>, finalFn: () => Promise<R>, reject: (err: Error) => void) {
-  return new Writable({
-    objectMode: true,
-    highWaterMark: params.highWaterMark ? params.highWaterMark || 4 : 4,
-    write(val: T, encoding: BufferEncoding, cb: (error?: Error | null, result?: any) => void) {
-      (async () => {
-        const start = process.hrtime();
-        const reqTime = start[0] * 1e3 + start[1] / 1e6;
-        const result = await processFn(val);
-        if ('mux' === params.name) {
-          const pktTimings = val.timings;
-          pktTimings[params.name] = { reqTime: reqTime, elapsed: process.hrtime(start)[1] / 1000000 };
-          if (doTimings)
-            timings.push(pktTimings);
-        }
-        cb(null, result);
-      })().catch(cb);
-    },
-    final(cb: (error?: Error | null, result?: any) => void) {
-      (async () => {
-        const result = finalFn ? await finalFn() : null;
-        if (doTimings && ('mux' === params.name)) {
-          const elapsedStats = {} as { [key: string]: ffStats };
-          Object.keys(timings[0]).forEach(k => elapsedStats[k] = calcStats<typeof k, 'elapsed'>(timings.slice(10, -10), k, 'elapsed'));
-          console.log('elapsed:');
-          console.table(elapsedStats);
-
-          const absArr = timings.map(t => {
-            const absDelays = {} as { [key: string]: { reqDelta: number } };
-            const keys = Object.keys(t);
-            keys.forEach((k, i) => absDelays[k] = { reqDelta: i > 0 ? t[k].reqTime - t[keys[i - 1]].reqTime : 0 });
-            return absDelays;
-          });
-          const absStats = {} as { [key: string]: ffStats };
-          Object.keys(absArr[0]).forEach(k => absStats[k] = calcStats<typeof k, 'reqDelta'>(absArr.slice(10, -10), k, 'reqDelta'));
-          console.log('request time delta:');
-          console.table(absStats);
-
-          const totalsArr = timings.map(t => {
-            const total = (t.mux && t.read) ? t.mux.reqTime - t.read.reqTime + t.mux.elapsed : 0;
-            return { total: { total: total } };
-          });
-          console.log('total time:');
-          console.table(calcStats<'total', 'total'>(totalsArr.slice(10, -10), 'total', 'total'));
-        }
-        cb(null, result);
-      })().catch(cb);
-    }
-  }).on('error', err => reject(err));
-}
-
-function readStream(params: { highWaterMark?: number }, demuxer: Demuxer, ms: { end: number }, index: number): Readable {
-  const time_base = demuxer.streams[index].time_base;
-  const end_pts = ms ? ms.end * time_base[1] / time_base[0] : Number.MAX_SAFE_INTEGER;
-  async function getPacket(): Promise<Packet | null> {
-    let packet: Packet = {} as Packet;
-    do { packet = await demuxer.read(); }
-    while (packet && packet.stream_index !== index);
-    return packet;
-  }
-
-  return new Readable({
-    objectMode: true,
-    highWaterMark: params.highWaterMark ? params.highWaterMark || 4 : 4,
-    read() {
-      (async () => {
-        const start = process.hrtime();
-        const reqTime = start[0] * 1e3 + start[1] / 1e6;
-        const packet = await getPacket();
-        if (packet && (packet.pts < end_pts)) {
-          packet.timings = {};
-          packet.timings.read = { reqTime: reqTime, elapsed: process.hrtime(start)[1] / 1000000 };
-          this.push(packet);
-        } else
-          this.push(null);
-      })();
-    }
-  });
-}
-
-function createBeamWritableStream(params: { highwaterMark?: number }, governor: Governor): Writable {
-  const beamStream = new Writable({
-    highWaterMark: params.highwaterMark || 16384,
-    write: (chunk, encoding, cb) => {
-      (async () => {
-        await governor.write(chunk);
-        cb();
-      })();
-    }
-  });
-  return beamStream;
-}
-
-export function demuxerStream(params: { highwaterMark?: number }): WritableDemuxerStream {
-  const governor = new beamcoder.governor({});
-  const stream = createBeamWritableStream(params, governor) as WritableDemuxerStream;
-  stream.on('finish', () => governor.finish());
-  stream.on('error', console.error);
-  stream.demuxer = (options: { governor?: Governor }) => {
-    options.governor = governor;
-    // delay initialisation of demuxer until stream has been written to - avoids lock-up
-    return new Promise(resolve => setTimeout(async () => resolve(await beamcoder.demuxer(options)), 20));
-  };
-  return stream;
-}
-
-function createBeamReadableStream(params: { highwaterMark?: number }, governor: Governor): Readable {
-  const beamStream = new Readable({
-    highWaterMark: params.highwaterMark || 16384,
-    read: size => {
-      (async () => {
-        const chunk = await governor.read(size);
-        if (0 === chunk.length)
-          beamStream.push(null);
-        else
-          beamStream.push(chunk);
-      })();
-    }
-  });
-  return beamStream;
-}
-
-export function muxerStream(params: { highwaterMark: number }): ReadableMuxerStream {
-  const governor = new beamcoder.governor({ highWaterMark: 1 });
-  const stream = createBeamReadableStream(params, governor) as ReadableMuxerStream;
-  stream.on('end', () => governor.finish());
-  stream.on('error', console.error);
-  stream.muxer = (options) => {
-    options = options || {};
-    options.governor = governor;
-    return beamcoder.muxer(options);
-  };
-  return stream;
-}
-
-export async function makeSources(params: BeamstreamParams): Promise<void> {
-  if (!params.video) params.video = [];
-  if (!params.audio) params.audio = [];
-  params.video.forEach(p => p.sources.forEach((src: BeamstreamSource) => {
-    if (src.input_stream) {
-      const demuxerStream = beamcoder.demuxerStream({ highwaterMark: 1024 });
-      src.input_stream.pipe(demuxerStream);
-      src.formatP = demuxerStream.demuxer({ iformat: src.iformat, options: src.options });
-    } else
-      src.formatP = beamcoder.demuxer({ url: src.url, iformat: src.iformat, options: src.options });
-  }));
-  params.audio.forEach(p => p.sources.forEach((src: BeamstreamSource) => {
-    if (src.input_stream) {
-      const demuxerStream = beamcoder.demuxerStream({ highwaterMark: 1024 });
-      src.input_stream.pipe(demuxerStream);
-      src.formatP = demuxerStream.demuxer({ iformat: src.iformat, options: src.options });
-    } else
-      src.formatP = beamcoder.demuxer({ url: src.url, iformat: src.iformat, options: src.options });
-  }));
-
-  for (const video of params.video) {
-    for (const src of video.sources) {
-      src.format = await src.formatP;
-      if (src.ms && !src.input_stream)
-        src.format.seek({ time: src.ms.start });
-      await src.formatP;
-    }
-  }
-
-  for (const audio of params.audio) {
-    for (const src of audio.sources) {
-      src.format = await src.formatP;
-      if (src.ms && !src.input_stream)
-        src.format.seek({ time: src.ms.start });
-      await src.formatP;
-    }
-  }
-
-  params.video.forEach((p: BeamstreamChannel) => p.sources.forEach(src =>
-    src.stream = readStream({ highWaterMark: 1 }, src.format, src.ms, src.streamIndex)));
-  params.audio.forEach((p: BeamstreamChannel) => p.sources.forEach(src =>
-    src.stream = readStream({ highWaterMark: 1 }, src.format, src.ms, src.streamIndex)));
-}
-
-function runStreams(
-  streamType: 'video' | 'audio',
-  sources: Array<BeamstreamSource>,
-  filterer: Filterer,
-  streams: Array<BeamstreamStream>,
-  mux: { writeFrame: (pkts: void | Packet) => void },
-  muxBalancer: serialBalancer): Promise<void> {
-  return new Promise<void>((resolve, reject) => {
-    if (!sources.length)
-      return resolve();
-
-    const timeBaseStream: Stream = sources[0].format.streams[sources[0].streamIndex];
-    const filterBalancer = parallelBalancer({ name: 'filterBalance', highWaterMark: 1 }, streamType, sources.length);
-
-    sources.forEach((src: BeamstreamSource, srcIndex: number) => {
-      const decStream = transformStream<Packet, Timable & Promise<DecodedFrames>>(
-        { name: 'decode', highWaterMark: 1 },
-        (pkts: Packet) => src.decoder.decode(pkts),
-        () => src.decoder.flush(), reject);
-
-      const filterSource = writeStream<DecodedFrames, localFrame>(
-        { name: 'filterSource', highWaterMark: 1 },
-        (pkts: DecodedFrames) => filterBalancer.pushPkts(pkts, src.format.streams[src.streamIndex], srcIndex, false),
-        () => filterBalancer.pushPkts(null, src.format.streams[src.streamIndex], srcIndex, true),
-        reject
-      );
-
-      src.stream.pipe(decStream).pipe(filterSource);
-    });
-
-    const streamTee = teeBalancer({ name: 'streamTee', highWaterMark: 1 }, streams.length);
-
-    const filtStream = transformStream<Timables<DecodedFrames>, Timable & Promise<Array<FiltererResult> & TotalTimeed>>({ name: 'filter', highWaterMark: 1 }, (frms: Timables<DecodedFrames>) => {
-      if (filterer.cb) filterer.cb(frms[0].frames[0].pts);
-      return filterer.filter(frms);
-    }, () => { }, reject);
-
-    const streamSource = writeStream<Timables<Frame>, BalanceResult | void>(
-      { name: 'streamSource', highWaterMark: 1 },
-      frms => streamTee.pushFrames(frms),
-      () => streamTee.pushFrames([], true),
-      reject
-    );
-
-    filterBalancer.pipe(filtStream).pipe(streamSource);
-    streams.forEach((str: BeamstreamStream, i: number) => {
-      const dicer = new frameDicer(str.encoder, 'audio' === streamType);
-
-      const diceStream = transformStream<Timables<Frame>, Timables<Frame>>(
-        { name: 'dice', highWaterMark: 1 },
-        (frms) => dicer.dice(frms),
-        () => dicer.dice([], true),
-        reject
-      );
-
-      const encStream = transformStream<Timables<Frame>, Timable & Promise<EncodedPackets>>(
-        { name: 'encode', highWaterMark: 1 },
-        (frms) => str.encoder.encode(frms),
-        () => str.encoder.flush(),
-        reject
-      );
-
-      const muxStream = writeStream<EncodedPackets, void | Packet>(
-        { name: 'mux', highWaterMark: 1 },
-        (pkts: EncodedPackets) => muxBalancer.writePkts(pkts, timeBaseStream, str.stream, pkts => mux.writeFrame(pkts)),
-        () => muxBalancer.writePkts(null, timeBaseStream, str.stream, pkts => mux.writeFrame(pkts), true),
-        reject
-      );
-
-      muxStream.on('finish', resolve);
-
-      streamTee[i].pipe(diceStream).pipe(encStream).pipe(muxStream);
-    });
-  });
-}
-
-export async function makeStreams(params: BeamstreamParams): Promise<{ run(): Promise<void> }> {
-  params.video.forEach((channel: BeamstreamChannel) => {
-    channel.sources.forEach((src: BeamstreamSource) =>
-      src.decoder = beamcoder.decoder({ demuxer: src.format, stream_index: src.streamIndex }));
-  });
-  params.audio.forEach((channel: BeamstreamChannel) => {
-    channel.sources.forEach((src: BeamstreamSource) =>
-      src.decoder = beamcoder.decoder({ demuxer: src.format, stream_index: src.streamIndex }));
-  });
-
-  const promises: Promise<Filterer>[] = [];
-
-  params.video.forEach((channel: BeamstreamChannel) => {
-    const inputParams = channel.sources.map((src, i) => {
-      const { codecpar, time_base, sample_aspect_ratio } = src.format.streams[src.streamIndex];
-      return {
-        name: `in${i}:v`,
-        width: codecpar.width,
-        height: codecpar.height,
-        pixelFormat: codecpar.format,
-        timeBase: time_base,
-        pixelAspect: sample_aspect_ratio
-      };
-    });
-
-    const p = beamcoder.filterer({
-      filterType: 'video',
-      inputParams,
-      outputParams: channel.streams.map((str: BeamstreamStream, i: number) => ({ name: `out${i}:v`, pixelFormat: str.codecpar.format })),
-      filterSpec: channel.filterSpec
-    }).then(filter => channel.filter = filter);
-    promises.push(p);
-  });
-
-  params.audio.forEach((channel: BeamstreamChannel) => {
-    const inputParams = channel.sources.map((src: BeamstreamSource, i: number) => {
-      const { sample_rate, sample_fmt, channel_layout } = src.decoder;
-      return {
-        name: `in${i}:a`,
-        sampleRate: sample_rate,
-        sampleFormat: sample_fmt,
-        channelLayout: channel_layout,
-        timeBase: src.format.streams[src.streamIndex].time_base
-      };
-    });
-    const outputParams = channel.streams.map((str: BeamstreamStream, i: number) => {
-      const { sample_rate, format, channel_layout } = str.codecpar;
-      return {
-        name: `out${i}:a`,
-        sampleRate: sample_rate,
-        sampleFormat: format,
-        channelLayout: channel_layout
-      };
-    });
-    const p = beamcoder.filterer({ filterType: 'audio', inputParams, outputParams, filterSpec: channel.filterSpec })
-    .then(filter => channel.filter = filter);
-    promises.push(p);
-  });
-  await Promise.all(promises);
-
-  // params.video.forEach(p => console.log(p.filter.graph.dump()));
-  // params.audio.forEach(p => console.log(p.filter.graph.dump()));
-
-  let mux: Muxer;
-  if (params.out.output_stream) {
-    let muxerStream = beamcoder.muxerStream({ highwaterMark: 1024 });
-    muxerStream.pipe(params.out.output_stream);
-    mux = muxerStream.muxer({ format_name: params.out.formatName });
-  } else {
-    mux = beamcoder.muxer({ format_name: params.out.formatName });
-  }
-  params.video.forEach((channel: BeamstreamChannel) => {
-    channel.streams.forEach((str: BeamstreamStream, i: number) => {
-      const encParams = channel.filter.graph.filters.find(f => f.name === `out${i}:v`).inputs[0];
-      str.encoder = beamcoder.encoder({
-        name: str.name,
-        width: encParams.w,
-        height: encParams.h,
-        pix_fmt: encParams.format,
-        sample_aspect_ratio: encParams.sample_aspect_ratio,
-        time_base: encParams.time_base as [number, number],
-        // framerate: [encParams.time_base[1], encParams.time_base[0]],
-        // bit_rate: 2000000,
-        // gop_size: 10,
-        // max_b_frames: 1,
-        // priv_data: { preset: 'slow' }
-        priv_data: { crf: 23 }
-      }); // ... more required ...
-    });
-  });
-
-  params.audio.forEach((channel: BeamstreamChannel) => {
-    channel.streams.forEach((str: BeamstreamStream, i: number) => {
-      const encParams: FilterLink = channel.filter.graph.filters.find(f => f.name === `out${i}:a`).inputs[0];
-      str.encoder = beamcoder.encoder({
-        name: str.name,
-        sample_fmt: encParams.format,
-        sample_rate: encParams.sample_rate,
-        channel_layout: encParams.channel_layout,
-        flags: { GLOBAL_HEADER: mux.oformat.flags.GLOBALHEADER }
-      });
-      str.codecpar.frame_size = str.encoder.frame_size;
-    });
-  });
-
-  params.video.forEach((channel: BeamstreamChannel) => {
-    channel.streams.forEach((str: BeamstreamStream) => {
-      str.stream = mux.newStream({
-        name: str.name,
-        time_base: str.time_base,
-        interleaved: true
-      }); // Set to false for manual interleaving, true for automatic
-      Object.assign(str.stream.codecpar, str.codecpar);
-    });
-  });
-
-  params.audio.forEach((channel: BeamstreamChannel) => {
-    channel.streams.forEach((str: BeamstreamStream) => {
-      str.stream = mux.newStream({
-        name: str.name,
-        time_base: str.time_base,
-        interleaved: true
-      }); // Set to false for manual interleaving, true for automatic
-      Object.assign(str.stream.codecpar, str.codecpar);
-    });
-  });
-
-  return {
-    run: async () => {
-      await mux.openIO({
-        url: params.out.url ? params.out.url : '',
-        flags: params.out.flags ? params.out.flags : {}
-      });
-      await mux.writeHeader({ options: params.out.options ? params.out.options : {} });
-
-      const muxBalancer = new serialBalancer(mux.streams.length);
-      const muxStreamPromises: Promise<void>[] = [];
-      params.video.forEach(p => muxStreamPromises.push(runStreams('video', p.sources, p.filter, p.streams, mux, muxBalancer)));
-      params.audio.forEach(p => muxStreamPromises.push(runStreams('audio', p.sources, p.filter, p.streams, mux, muxBalancer)));
-      await Promise.all(muxStreamPromises);
-
-      await mux.writeTrailer();
-    }
-  };
-}
