+++ ts/demuxerStream.ts
@@ -2,7 +2,7 @@
   Aerostat Beam Coder - Node.js native bindings to FFmpeg
   Copyright (C) 2019 Streampunk Media Ltd.
   Copyright (C) 2022 Chemouni Uriel.
-  
+
   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
@@ -19,8 +19,22 @@
   https://www.streampunk.media/ mailto:furnace@streampunk.media
   14 Ormiscaig, Aultbea, Achnasheen, IV22 2JJ  U.K.
 */
-import bindings from 'bindings';
-import { BeamcoderType } from './types/BeamcoderType';
+import type { Governor } from './types/Governor';
+import { WritableDemuxerStream } from './types/Beamstreams';
 
-const beamcoder = bindings('beamcoder') as BeamcoderType;
-export default beamcoder;
+import beamcoder from './beamcoder'
+import createBeamWritableStream from './createBeamWritableStream';
+
+export default function demuxerStream(params: { highwaterMark?: number }): WritableDemuxerStream {
+    const governor = new beamcoder.governor({});
+    const stream = createBeamWritableStream(params, governor) as WritableDemuxerStream;
+    stream.on('finish', () => governor.finish());
+    stream.on('error', console.error);
+    stream.demuxer = (options: { governor?: Governor }) => {
+      options.governor = governor;
+      // delay initialisation of demuxer until stream has been written to - avoids lock-up
+      return new Promise(resolve => setTimeout(async () => resolve(await beamcoder.demuxer(options)), 20));
+    };
+    return stream;
+  }
+
