+++ ts/install_ffmpeg.ts
@@ -1,6 +1,7 @@
 /*
   Aerostat Beam Coder - Node.js native bindings to FFmpeg.
   Copyright (C) 2019  Streampunk Media Ltd.
+  Copyright (C) 2022 Chemouni Uriel.
 
   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
