+++ tsconfig.json
@@ -98,5 +98,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
   },
-  "files": [ "ts/beamstreams.ts", "ts/index.ts", "ts/install_ffmpeg.ts" ]
+  "include": ["ts/*.ts"],
 }
